Collect docnames of analyzed source in advance (#2)
* Remove container_of and support "of {key: value}" syntax

This syntax is used by Pandas [1]. I'm not entirely sold yet, but let's
add it for now. To avoid confusion, literals should probably be made to
only be accepted on the top-level. Otherwise something like
`dict of {{"a", "b"}: int}` becomes possible.

Co-authored-by: Oriol Abril-Pla <oriol.abril.pla@gmail.com>

* Only allow literals on top-level

which avoids potentially confusing constructs like
`dict of {{"a", "b"}: int}`

Co-authored-by: Oriol Abril-Pla <oriol.abril.pla@gmail.com>

* Make "= | :" optional in default syntax

since this is what NumPyDoc recommends as well [1].

[1] https://numpydoc.readthedocs.io/en/latest/format.html#parameters

Co-authored-by: Oriol Abril-Pla <oriol.abril.pla@gmail.com>

* Add ipython as dev dependency

* Move matching to imports to qualname level

Doing this on the NAME level meant that stuff like "np.int16" would be
turned into "np.np.int16" because both "np" and "int16" where matched.

* Split docname into KnownImport and replace

Reduce responsibility of the former DocName class. Replacing docstring
specific type description should be handled separately.

* Test correctness when nesting classes in classes

* Rework analysis and other major changes

While not perfect or anywhere near to finished the refactored code
uses clearer separation of responsiblities and is one step closer to an
architecture that feels right. :)

---------

Co-authored-by: Oriol Abril-Pla <oriol.abril.pla@gmail.com>