Only use `| None` for optional parameters if appropriate (#14)
* Allow sphinx references enclosed in "`" only

The prefix in ":class:`Foo`" is actually optional for Sphinx to detect
it as a reference. So allow "`Foo`" too.

* Only use `| None` for optional parameters if appropriate

This commit also includes a few other tweaks. I should really pay a bit
more attention to making more focused changes. Partial commits only get
you so far...

A minor change is that `PackageFile` was replaced with
`module_name_from_path` which should make for a more sensible approach
that doesn't rely as much on carrying a subclass of `Path` around.