Support attributes and type aliases (#18)
* Support collection of explicit type aliases

While it is very difficult to statically detect a type alias [1] such as
`x = str`, explicit ones are easier to detect. This change makes sure
that `x: TypeAlias = str` and `type x = str` are collected as importable
type annotations.

[1] https://docs.python.org/3/library/typing.html#type-aliases

* Handle attributes in stub generation

Type information in docstrings is not yet taken into account. Instead,
`typing.Any` is used as a fallback.

* Use `_typeshed.Incomplete` as a fallback type instead

* Refactor enum `FuncType` to `ScopeType`

* Add support for attribute annotation with doctypes

as well as proper handling in context of stubs. In stubs, the value
should basically always be omitted from attributes except for a few
exceptions.

* Handle attributes in stub generation

Type information in docstrings is not yet taken into account. Instead,
`typing.Any` is used as a fallback.

* Blacklist reserved Python keywords from qualnames