Add docstub comment directives to toggle transformation
This enables using `# docstub: off` and `# docstub: on` comments – a way
to tell docstub to copy wrapped code unchanged into a stub file. Not
sure yet how useful this will be, but for now it's self-contained enough
and easy to remove.

What would be really nice, would be a way to override the output of
docstub in a similar manner – in case docstub doesn't do the right thing
or doesn't yet know how to.

E.g. to handle edge cases like attributes in enums etc.