Fix grammar and allow plural "(s)" (#37)
Removes an implicit match for "" (nothing) because I left the first line
after the ":" empty. Obviously not what is intended! So while it would
be more readable to be able to wrap everything inside a rule with an
indent, it doesn't seem to be supported currently [1].

Also, allow appending `(s)` to indicate plural in expressions like
`list of int(s)`. A negative lookbehind regex disallows a space directly
in front of `(s)`.

Other changes improve on naming, structure and tests.

[1] www.github.com/lark-parser/lark/issues/155