Add minimal documentation & tweaks (#24)
Adds a user guide, a reference for the typing syntax

Adds a user guide, a reference for the typing syntax and the command
line as well as tests that verify parts of it.

In addition, this ended up incorporating a bigger & smaller tweaks that
refactor docstub and change behavior. These came up during documentation
when I noticed behavior that was hard to document or didn't seem very
intuitive.

Co-authored-by: Lars Grüter <lagru@mailbox.org>
Co-authored-by: Oriol Abril-Pla <oriol.abril.pla@gmail.com>