Keep inline annotations over docstring generated ones (#61)
* first pass at keeping inline annotation over docstring

* Make override_docsring tests more independent

Keeping parameters is already tested in `override_docstring_param`
above – no need to do so again.

* Remove duplication in error message

The `underline` method of the `ErrorReporter` includes the underlined
line already.

* Simplify logic when ignoring docstring annotation

Also makes sure that imports of a doctype are not unnecessarily
included an inline annotation already exists.

* Keep existing inline annotation for assignments

even if a different docstring annotation exists. This is the same
behavior as for parameters and return types.

* Add more tests

* Re-enable commented out tests

Probably forgot to do so during debugging.

* Document inline annotation as overwrite solution

---------

Co-authored-by: Lars Grüter <lagru@mailbox.org>