Improve handling of inline annotations without conflicts (#66)
Make sure that no warning is raised for inline annotations, if there is no
conflicting annotation in a docstring. Otherwise, the inlined annotation 
takes precedence and a warning is printed.

* Don't raise note for inline annotated attributes

when the docstring doesn't contain a type description. In which case
a `FallbackAnnotation` is provided.

Also simplifies the redundant implementation of
`DocstringAnnotations.attributes`.

* Improve tests for doctype-inline conflicts