Join "optional" and "extra_info" rules in grammar
Since the "optional" statement isn't used in any way right now,
enforcing a specific layout for what comes after the "doctype" serves
no useful purpose. I don't intend docstub to be linter so I don't think
it's likely that we will ever enforce annotating that a parameter is
"optional" in the docstring – while it's nice to do so, it's reasonably
easy to see it from the signature.

Update the docs accordingly. Also try to be more precise about
differentiating "doctypes" from valid Python types.