Add logging-based error reporting (#79)
Improve output and error reporting. All output of docstub is now controllable
with the existing `--verbose` option and the new `--quiet` option. Also indicate
the (severity) level of output with a single char, e.g. "I" for information or
"W" for warnings.

Refactor reporting to use Python's native logging implementation.