Track generated inplace stubs for docstub itself
I thought about not doing this but then it becomes more complicated
to include the stubs into the package in the release workflow (cd.yml).
I want to keep the release workflow as minimal as possible to reduce
the risk of a supply chain attack.

This approach sidesteps that problem, because the stubs are already
there and tracked. However, we need to assert that the tracked stubs
are identical with the docstub generated ones. That is a bit tricky,
because `git diff` actually ignores untracked files! So I came up with
`assert-unchanged.sh`. I really struggled with this becaue I initally
tried to use git ls-files with the "-z" option which returns
null-delimited paths. I couldn't figure out how to pass them to
`git add` and `git unstage` later on.

I'm worried that this approach is a bit brittle (due to
`assert-unchanged.sh`) and increases the diff in PRs. On the other hand
being able to review the stubs is probably not a bad thing. We'll see
how that goes.