Add `--workers` option controlling parallel stub generation (#110)
Add the new experimental `--workers` option to `docstub run` to generate stub files
in parallel. This can reduce runtime significantly for projects with many files.
For now, multiprocessing has to be explicitly enabled by requesting a number of
workers.

Also ensure that docstub still returns a non-zero exit code when it is run with output
completely disabled (`-qq`). The same applies when running docstub with 
`--quiet --fail-on-warnings`.

--

* Draft multiprocessing with PoolExecutor

* Count errors in dedicated handler

Splitting these two responsibilities (logging to console and counting)
simplifies both in context of multiprocessing.

* Update how stats are propagated back to the main process

* Add critical method to ContextReporter

* Use "--workers" instead of "--jobs"

The former is what scikit-image agreed on.

* Test error and warning propagation with -qq

Mark these as "slow" (because they are)

* Add reminders to remove vendored stdlib code eventually

* Make docstest independent of available CPUs

* Add tests for `guess_concurrency_params`

* Fix swallowing exit code for diffs

* Improve exit function of LoggingProcessExecutor

* Fix stubtest errors

* Update example_pkg-stubs

* Sync _concurrency.pyi

* Mark --worker option as experimental and default to single process

This is probably safest for now. I don't want to break peoples setup
the new version of docstub suddenly hangs or fails. This way they users
have to opt in explicitly and will now what might have caused it.

* Fix typo in docstest caused by accidental refactor

* Test multiprocessing in CI

* Factor CI multiprocessing step into its own step

* Fix test setup for monkeypatching process_cpu_count

* Skip subprocess tests on Linux 3.12-3.13

* Start logging queue is in same multiprocessing context as pool

This fixes the following:

    RuntimeError: A SemLock created in a fork context is being shared
    with a process in a spawn context. This is not supported. Please use
    the same context to create multiprocessing objects and Process.

* Fix unkown "Queue" in doctype